Zerodha_algo/requirements.txt
kiteconnect>=2.3.3
pandas
numpy
streamlit
python-dotenv
pytest